+++ Process.c
@@ -793,13 +793,13 @@ void Process_sendSignal(Process* this, size_t sgn) {
    kill(this->pid, (int) sgn);
 }
 
-int Process_pidCompare(const void* v1, const void* v2) {
+long Process_pidCompare(const void* v1, const void* v2) {
    Process* p1 = (Process*)v1;
    Process* p2 = (Process*)v2;
    return (p1->pid - p2->pid);
 }
 
-int Process_compare(const void* v1, const void* v2) {
+long Process_compare(const void* v1, const void* v2) {
    Process *p1, *p2;
    ProcessList *pl = ((Process*)v1)->pl;
    if (pl->direction == 1) {
